# src/utils/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "School-ERP APIs",
    "version": "1.0.0",
    "description": "Comprehensive School Management System API with multi-tenant architecture"
  },
...
  "paths":{
...
"/school/{id}": {
      "put": {
        "tags": ["School"],
        "summary": "Update school",
        "description": "Update details of a specific school",
        "operationId": "updateSchool",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID of the school"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "institutionName": { "type": "string" },
                  "institutionType": { "type": "string" },
                  "board": { "type": "string" },
                  "establishmentYear": { "type": "integer" },
                  "medium": { "type": "string" },
                  "totalStudent": { "type": "integer" },
                  "contactPersonName": { "type": "string" },
                  "contactNumber": { "type": "string" },
                  "alternateNumber": { "type": "string" },
                  "addressLine1": { "type": "string" },
                  "addressLine2": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "city": { "type": "string" },
                  "state": { "type": "string" },
                  "pincode": { "type": "string" },
                  "country": { "type": "string" },
                  "plan": { "type": "string" },
                  "smsSenderId": { "type": "string" },
                  "whatsappNumber": { "type": "string" },
                  "adminName": { "type": "string" },
                  "adminEmail": { "type": "string", "format": "email" },
                  "adminMobileNumber": { "type": "string" },
                  "notes": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "School updated successfully"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "School not found"
          }
        }
      },
      "delete": {
        "tags": ["School"],
        "summary": "Delete school",
        "description": "Deletes a school and related admin/schema",
        "operationId": "deleteSchool",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID of the school to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "School and its related data deleted successfully"
          },
          "404": {
            "description": "School not found"
          },
          "500": {
            "description": "Error during deletion"
          }
        }
      },
      "get": {
      "tags": ["School"],
      "summary": "Get a single school by ID",
      "description": "Retrieves a school by its ID including its plan",
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "responses": {
        "200": {
          "description": "School retrieved successfully"
        },
        "400": {
          "description": "School ID is required"
        },
        "500": {
          "description": "Server error"
        }
      }
    }
  },
...
}}